Factor input latency reporting into its own crate and not in GPUI
